@@ -1,19 +1,20 @@
 <?php
+
 /**
  * SCSSPHP
  *
- * @copyright 2012-2018 Leaf Corcoran
+ * @copyright 2012-2020 Leaf Corcoran
  *
  * @license http://opensource.org/licenses/MIT MIT
  *
- * @link http://leafo.github.io/scssphp
+ * @link http://scssphp.github.io/scssphp
  */
 
-namespace Leafo\ScssPhp\Node;
+namespace ScssPhp\ScssPhp\Node;
 
-use Leafo\ScssPhp\Compiler;
-use Leafo\ScssPhp\Node;
-use Leafo\ScssPhp\Type;
+use ScssPhp\ScssPhp\Compiler;
+use ScssPhp\ScssPhp\Node;
+use ScssPhp\ScssPhp\Type;
 
 /**
  * Dimension + optional units
@@ -28,17 +29,20 @@ use Leafo\ScssPhp\Type;
  */
 class Number extends Node implements \ArrayAccess
 {
+    const PRECISION = 10;
+
     /**
      * @var integer
+     * @deprecated use {Number::PRECISION} instead to read the precision. Configuring it is not supported anymore.
      */
-    static public $precision = 10;
+    public static $precision = self::PRECISION;
 
     /**
      * @see http://www.w3.org/TR/2012/WD-css3-values-20120308/
      *
      * @var array
      */
-    static protected $unitTable = [
+    protected static $unitTable = [
         'in' => [
             'in' => 1,
             'pc' => 6,
@@ -64,8 +68,8 @@ class Number extends Node implements \ArrayAccess
         ],
         'dpi' => [
             'dpi'  => 1,
-            'dpcm' => 2.54,
-            'dppx' => 96,
+            'dpcm' => 1 / 2.54,
+            'dppx' => 1 / 96,
         ],
     ];
 
@@ -89,7 +93,7 @@ class Number extends Node implements \ArrayAccess
     {
         $this->type      = Type::T_NUMBER;
         $this->dimension = $dimension;
-        $this->units     = is_array($initialUnit)
+        $this->units     = \is_array($initialUnit)
             ? $initialUnit
             : ($initialUnit ? [$initialUnit => 1]
                             : []);
@@ -100,7 +104,7 @@ class Number extends Node implements \ArrayAccess
      *
      * @param array $units
      *
-     * @return \Leafo\ScssPhp\Node\Number
+     * @return \ScssPhp\ScssPhp\Node\Number
      */
     public function coerce($units)
     {
@@ -110,11 +114,18 @@ class Number extends Node implements \ArrayAccess
 
         $dimension = $this->dimension;
 
-        foreach (static::$unitTable['in'] as $unit => $conv) {
-            $from       = isset($this->units[$unit]) ? $this->units[$unit] : 0;
-            $to         = isset($units[$unit]) ? $units[$unit] : 0;
-            $factor     = pow($conv, $from - $to);
-            $dimension /= $factor;
+        if (\count($units)) {
+            $baseUnit = array_keys($units);
+            $baseUnit = reset($baseUnit);
+            $baseUnit = $this->findBaseUnit($baseUnit);
+            if ($baseUnit && isset(static::$unitTable[$baseUnit])) {
+                foreach (static::$unitTable[$baseUnit] as $unit => $conv) {
+                    $from       = isset($this->units[$unit]) ? $this->units[$unit] : 0;
+                    $to         = isset($units[$unit]) ? $units[$unit] : 0;
+                    $factor     = pow($conv, $from - $to);
+                    $dimension /= $factor;
+                }
+            }
         }
 
         return new Number($dimension, $units);
@@ -123,14 +134,14 @@ class Number extends Node implements \ArrayAccess
     /**
      * Normalize number
      *
-     * @return \Leafo\ScssPhp\Node\Number
+     * @return \ScssPhp\ScssPhp\Node\Number
      */
     public function normalize()
     {
         $dimension = $this->dimension;
         $units     = [];
 
-        $this->normalizeUnits($dimension, $units, 'in');
+        $this->normalizeUnits($dimension, $units);
 
         return new Number($dimension, $units);
     }
@@ -141,17 +152,18 @@ class Number extends Node implements \ArrayAccess
     public function offsetExists($offset)
     {
         if ($offset === -3) {
-            return $this->sourceColumn !== null;
+            return ! \is_null($this->sourceColumn);
         }
 
         if ($offset === -2) {
-            return $this->sourceLine !== null;
+            return ! \is_null($this->sourceLine);
         }
 
-        if ($offset === -1
-            || $offset === 0
-            || $offset === 1
-            || $offset === 2
+        if (
+            $offset === -1 ||
+            $offset === 0 ||
+            $offset === 1 ||
+            $offset === 2
         ) {
             return true;
         }
@@ -232,6 +244,35 @@ class Number extends Node implements \ArrayAccess
     }
 
     /**
+     * Test if a number can be normalized in a base unit
+     * ie if its units are homogeneous
+     *
+     * @return boolean
+     */
+    public function isNormalizable()
+    {
+        if ($this->unitless()) {
+            return false;
+        }
+
+        $baseUnit = null;
+
+        foreach ($this->units as $unit => $exp) {
+            $b = $this->findBaseUnit($unit);
+
+            if (\is_null($baseUnit)) {
+                $baseUnit = $b;
+            }
+
+            if (\is_null($b) or $b !== $baseUnit) {
+                return false;
+            }
+        }
+
+        return $baseUnit;
+    }
+
+    /**
      * Returns unit(s) as the product of numerator units divided by the product of denominator units
      *
      * @return string
@@ -243,41 +284,41 @@ class Number extends Node implements \ArrayAccess
 
         foreach ($this->units as $unit => $unitSize) {
             if ($unitSize > 0) {
-                $numerators = array_pad($numerators, count($numerators) + $unitSize, $unit);
+                $numerators = array_pad($numerators, \count($numerators) + $unitSize, $unit);
                 continue;
             }
 
             if ($unitSize < 0) {
-                $denominators = array_pad($denominators, count($denominators) + $unitSize, $unit);
+                $denominators = array_pad($denominators, \count($denominators) - $unitSize, $unit);
                 continue;
             }
         }
 
-        return implode('*', $numerators) . (count($denominators) ? '/' . implode('*', $denominators) : '');
+        return implode('*', $numerators) . (\count($denominators) ? '/' . implode('*', $denominators) : '');
     }
 
     /**
      * Output number
      *
-     * @param \Leafo\ScssPhp\Compiler $compiler
+     * @param \ScssPhp\ScssPhp\Compiler $compiler
      *
      * @return string
      */
     public function output(Compiler $compiler = null)
     {
-        $dimension = round($this->dimension, static::$precision);
+        $dimension = round($this->dimension, self::PRECISION);
 
         $units = array_filter($this->units, function ($unitSize) {
             return $unitSize;
         });
 
-        if (count($units) > 1 && array_sum($units) === 0) {
+        if (\count($units) > 1 && array_sum($units) === 0) {
             $dimension = $this->dimension;
             $units     = [];
 
-            $this->normalizeUnits($dimension, $units, 'in');
+            $this->normalizeUnits($dimension, $units);
 
-            $dimension = round($dimension, static::$precision);
+            $dimension = round($dimension, self::PRECISION);
             $units     = array_filter($units, function ($unitSize) {
                 return $unitSize;
             });
@@ -285,15 +326,17 @@ class Number extends Node implements \ArrayAccess
 
         $unitSize = array_sum($units);
 
-        if ($compiler && ($unitSize > 1 || $unitSize < 0 || count($units) > 1)) {
-            $compiler->throwError((string) $dimension . $this->unitStr() . " isn't a valid CSS value.");
+        if ($compiler && ($unitSize > 1 || $unitSize < 0 || \count($units) > 1)) {
+            $this->units = $units;
+            $unit = $this->unitStr();
+        } else {
+            reset($units);
+            $unit = key($units);
         }
 
-        reset($units);
-        $unit = key($units);
-        $dimension = number_format($dimension, static::$precision, '.', '');
+        $dimension = number_format($dimension, self::PRECISION, '.', '');
 
-        return (static::$precision ? rtrim(rtrim($dimension, '0'), '.') : $dimension) . $unit;
+        return (self::PRECISION ? rtrim(rtrim($dimension, '0'), '.') : $dimension) . $unit;
     }
 
     /**
@@ -311,13 +354,17 @@ class Number extends Node implements \ArrayAccess
      * @param array         $units
      * @param string        $baseUnit
      */
-    private function normalizeUnits(&$dimension, &$units, $baseUnit = 'in')
+    private function normalizeUnits(&$dimension, &$units, $baseUnit = null)
     {
         $dimension = $this->dimension;
         $units     = [];
 
         foreach ($this->units as $unit => $exp) {
-            if (isset(static::$unitTable[$baseUnit][$unit])) {
+            if (! $baseUnit) {
+                $baseUnit = $this->findBaseUnit($unit);
+            }
+
+            if ($baseUnit && isset(static::$unitTable[$baseUnit][$unit])) {
                 $factor = pow(static::$unitTable[$baseUnit][$unit], $exp);
 
                 $unit = $baseUnit;
@@ -326,5 +373,23 @@ class Number extends Node implements \ArrayAccess
 
             $units[$unit] = $exp + (isset($units[$unit]) ? $units[$unit] : 0);
         }
+    }
+
+    /**
+     * Find the base unit family for a given unit
+     *
+     * @param string $unit
+     *
+     * @return string|null
+     */
+    private function findBaseUnit($unit)
+    {
+        foreach (static::$unitTable as $baseUnit => $unitVariants) {
+            if (isset($unitVariants[$unit])) {
+                return $baseUnit;
+            }
+        }
+
+        return null;
     }
 }
