@@ -1,15 +1,16 @@
 <?php
+
 /**
  * SCSSPHP
  *
- * @copyright 2012-2018 Leaf Corcoran
+ * @copyright 2012-2020 Leaf Corcoran
  *
  * @license http://opensource.org/licenses/MIT MIT
  *
- * @link http://leafo.github.io/scssphp
+ * @link http://scssphp.github.io/scssphp
  */
 
-namespace Leafo\ScssPhp;
+namespace ScssPhp\ScssPhp;
 
 /**
  * Block
@@ -24,7 +25,7 @@ class Block
     public $type;
 
     /**
-     * @var \Leafo\ScssPhp\Block
+     * @var \ScssPhp\ScssPhp\Block
      */
     public $parent;
 
@@ -62,4 +63,9 @@ class Block
      * @var array
      */
     public $children;
+
+    /**
+     * @var \ScssPhp\ScssPhp\Block
+     */
+    public $selfParent;
 }
