@@ -1,15 +1,16 @@
 <?php
+
 /**
  * SCSSPHP
  *
- * @copyright 2012-2018 Leaf Corcoran
+ * @copyright 2012-2020 Leaf Corcoran
  *
  * @license http://opensource.org/licenses/MIT MIT
  *
- * @link http://leafo.github.io/scssphp
+ * @link http://scssphp.github.io/scssphp
  */
 
-namespace Leafo\ScssPhp\Compiler;
+namespace ScssPhp\ScssPhp\Compiler;
 
 /**
  * Compiler environment
@@ -19,12 +20,12 @@ namespace Leafo\ScssPhp\Compiler;
 class Environment
 {
     /**
-     * @var \Leafo\ScssPhp\Block
+     * @var \ScssPhp\ScssPhp\Block
      */
     public $block;
 
     /**
-     * @var \Leafo\ScssPhp\Compiler\Environment
+     * @var \ScssPhp\ScssPhp\Compiler\Environment
      */
     public $parent;
 
@@ -32,6 +33,11 @@ class Environment
      * @var array
      */
     public $store;
+
+    /**
+     * @var array
+     */
+    public $storeUnreduced;
 
     /**
      * @var integer
