@@ -1,15 +1,16 @@
 <?php
+
 /**
  * SCSSPHP
  *
- * @copyright 2012-2018 Leaf Corcoran
+ * @copyright 2012-2020 Leaf Corcoran
  *
  * @license http://opensource.org/licenses/MIT MIT
  *
- * @link http://leafo.github.io/scssphp
+ * @link http://scssphp.github.io/scssphp
  */
 
-namespace Leafo\ScssPhp;
+namespace ScssPhp\ScssPhp;
 
 /**
  * Block/node types
@@ -27,6 +28,7 @@ class Type
     const T_COMMENT = 'comment';
     const T_CONTINUE = 'continue';
     const T_CONTROL = 'control';
+    const T_CUSTOM_PROPERTY = 'custom';
     const T_DEBUG = 'debug';
     const T_DIRECTIVE = 'directive';
     const T_EACH = 'each';
@@ -37,6 +39,7 @@ class Type
     const T_EXTEND = 'extend';
     const T_FOR = 'for';
     const T_FUNCTION = 'function';
+    const T_FUNCTION_REFERENCE = 'function-reference';
     const T_FUNCTION_CALL = 'fncall';
     const T_HSL = 'hsl';
     const T_IF = 'if';
